utils and midi

+++ src/utils/stack.c
@@ -1,6 +1,7 @@
 #include <stdint.h>
 #include <stdio.h>
 
+#include "cell.h"
 #include "stack.h"
 
 Stack stack(void) {
@@ -16,10 +17,10 @@ void pushP(Stack *s, void *p) {
   s->cells[++s->n].p = p;
 }
 
-void pushI(Stack *s, int32_t n) {
+void pushI(Stack *s, int64_t n) {
   s->cells[++s->n].i = n; 
 }
 
-void pushF(Stack *s, float f) {
+void pushF(Stack *s, double f) {
   s->cells[++s->n].f = f;
 }
